backend/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
pydantic==2.9.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
boto3==1.34.162
minio==7.2.8
celery==5.4.0
redis==5.0.8
python-multipart==0.0.9
aiofiles==24.1.0
langchain==0.2.16
tika==2.6.0
nltk==3.9.1
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
